+++ NewPlatform.Flexberry.Analytics/Abstractions/IReportManager.cs
@@ -32,6 +32,6 @@
         /// <param name="reportPath">Путь к отчёту.</param>
         /// <param name="parameters">Параметры отчёта.</param>
         /// <param name="ct">Маркер отмены.</param>
-        Task<HttpResponseMessage> ExportReport(string reportPath, JObject parameters, CancellationToken ct);
+        Task<byte[]> ExportReport(string reportPath, JObject parameters, CancellationToken ct);
     }
 }