+++ NewPlatform.Flexberry.Analytics/WebApiSample/App_Start/WebApiConfig.cs
@@ -1,14 +1,13 @@
 ﻿namespace NewPlatform.Flexberry.Analytics.WebApiSample
 {
-    using NewPlatform.Flexberry.AspNet.WebApi.Cors;
     using System.Web.Http;
+    using NewPlatform.Flexberry.AspNet.WebApi.Cors;
 
     public static class WebApiConfig
     {
         public static void Register(HttpConfiguration config)
         {
             // Конфигурация и службы веб-API
-  
             config.EnableCors(new DynamicCorsPolicyProvider(true));
 
             // Маршруты веб-API
@@ -17,8 +16,7 @@
             config.Routes.MapHttpRoute(
                 name: "DefaultApi",
                 routeTemplate: "api/{controller}/{id}",
-                defaults: new { id = RouteParameter.Optional }
-            );
+                defaults: new { id = RouteParameter.Optional });
         }
     }
 }