+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [2.0.0-beta01]
 
 ### Added
+* Netstandard2.0 support
+* ASP.NET Core sample project
 
 ### Changed
+* IReportManager.ExportReport return type to byte[]
+* Updated Newtonsoft.Json to 13.0.1
 
 ### Deprecated